Placeholder
Idea is that you enter a person and an item and it tells you if that person will like that item as a gift or not. This is based on the [stardew wiki](https://stardewvalleywiki.com/List_of_All_Gifts) and is just a matter of transferring all of that to a site. I'm tired.